pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = ">=2.10.6,<2.11.0"
pydantic-settings = ">=2.8.1,<2.9.0"
fastapi = ">=0.115.9,<0.116.0"
python-multipart = ">=0.0.20,<0.1.0"
httpx = ">=0.28.1,<0.29.0"
alembic = ">=1.14.1,<1.15.0"
sqlmodel = ">=0.0.22,<0.1.0"
asyncpg = ">=0.30.0,<0.31.0"
uvicorn = ">=0.34.0,<0.35.0"
celery = ">=5.4.0,<5.5.0"
redis = ">=5.2.1,<5.3.0"
langchain = ">=0.3.19,<0.4.0"
langchain-chroma = ">=0.2.2,<0.3.0"
langchain-openai = ">=0.3.7,<0.4.0"
langchain-community = ">=0.3.18,<0.4.0"
chromadb = ">=0.6.3,<0.7.0"
baml-py = ">=0.77.0,<0.78.0"
spacy = ">=3.7.5,<3.8.0"
poetry-core = ">=2.1.0,<2.2.0"
structlog = ">=25.1.0,<25.2.0"
aiofiles = ">=24.1.0,<24.2.0"
pypdf = ">=5.3.0,<5.4.0"
python-docx = ">=1.1.2,<1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.6"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-httpx = "^0.35.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core>=2.1.0"]
build-backend = "poetry.core.masonry.api"
